+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/common
 
+## 1.2.3
+
+### Patch Changes
+
+- feat: support interruptting the entire workflow if one task throw one exception
+
 ## 1.2.2
 
 ### Patch Changes